* fix(install-drupal.sh): Check presence of index.html before renaming it

- If  someone runs the deployer more than once then this prevents the
- error `mv: cannot stat 'index.html': No such file or directory` from happening

* fix(install.sh): Prevent error output to from appended to oneclick-website-deployer.log

- A bit unsure if it is the correct way though

* fix(install-apache.sh): Check if symlink exists before creating it

- To prevent error 'ln: failed to create symbolic link './rewrite.load': File exists' from happening

* fix(setup-cron.sh): Initialise variable drupal_cron_key

- Somehow it was used to setup cronjob but not initialised anywhere

* feat(tripal-daemon): Add hourly cronjob to restart daemon

* docs(install-drupal.sh): Add TODO comments

+++ scripts/install-drupal.sh
@@ -8,15 +8,15 @@ fi
 # Install Drupal
 sudo chown -R "$USER" "$WEB_ROOT"
 cd "$WEB_ROOT" || exit
-mv index.html index.html.orig
+[ -f index.html ] && mv index.html index.html.orig
 drush dl drupal --drupal-project-rename="$DRUPAL_HOME"
 
 # Setup Drupal
 cd "$DRUPAL_HOME" || exit
 cp sites/default/default.settings.php sites/default/settings.php
 mkdir sites/default/files/
-sudo chgrp -R www-data sites/default/files/
-sudo chmod g+rw sites/default/files/
+sudo chgrp -R www-data sites/default/files/ # NOTE: Incorporate apachectl -S 2> /dev/null | grep 'User' | awk -F '"' '{print $2}'
+sudo chmod g+rw sites/default/files/ # TODO: Checkout Tripal 4 readthedocs and checkout other way to fix directory permissions
 
 drush site-install standard --db-url=pgsql://"$psqluser":"$PGPASSWORD"@localhost:5432/"$psqldb" --account-mail="$drupal_mail" --account-name="$drupal_user" --account-pass="$drupal_pass" --site-mail="$drupal_mail" --site-name="$drupal_site_name" install_configure_form.site_default_country="$drupal_country_code" -y
